+++ SkyLibrary/src/main/java/jme3utilities/sky/Constants.java
@@ -106,6 +106,6 @@ final public class Constants {
      * @return branch and revision (not null, not empty)
      */
     public static String versionShort() {
-        return "master 1.4.0";
+        return "master 1.4.1";
     }
 }